pyproject.toml
[tool.poetry]
name = "udata-event-service"
version = "0.0.8"
description = "udata event service package containing kafka utils"
authors = ["Opendatateam <opendatateam@data.gouv.fr>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
